+++ Beam/app/Console/Commands/AggregatePageviewLoadJob.php
@@ -26,7 +26,7 @@ class AggregatePageviewLoadJob extends Command
         $now = $this->option('now') ? Carbon::parse($this->option('now')) : Carbon::now();
 
         $timeBefore = $now->minute(0)->second(0)->microsecond(0);
-        $timeAfter = (clone $timeBefore)->subHour();
+        $timeAfter = (clone $timeBefore)->subMinutes(20);
 
         $this->line(sprintf("Fetching aggregated pageviews data from <info>%s</info> to <info>%s</info>.", $timeAfter, $timeBefore));